Set up Gatsby.js
This configures a basic Gatsby set-up to replicate our static index. It
doesn't yet assume any framework for styles. It also doesn't add the
Netlify CMS yet so it can still run on GitHub Pages without having a
Netlify account.

Data set-up is kept to a minimum and should be done in future commits.