Add about and team page
This sets up Gatsby so that we can add teammembers that are
automatically displayed on the about page. Additionally Gatsby is
configured to render pages from the pages folder. This allows the code
of conduct and acessibility statement to be easily added and modified.

Some quick and dirty markup is made with Tailwind to put everything in
place.

It's chosen to use a modal dialog to provide more information about the
user. In the Barcelona website this was hidden on a new page. Ideally we
increase the number of teammembers per row and make the page responsive
by collapsing to a single column on smaller screens.

A keyboard handler needs to be added for the modal to be a11y compliant.
A header is added to allow navigation between the homepage and the about
page. The color is moved there and removed from the rest of the page.
More design will be needed to spruce up those pages.