Create splash front-page with responsive menu
This commit splits the main layout into multiple different layouts. This
allows a splash page to be created without visisble header. It also
allows the default pages to be created with visible header.

This commit adds a SkipLink to be able to skip the menu using a
screenreader or keybaord.

This commit does not yet put the logo and logo text back in the menu.