Refactor MainNavigation to re-usable Header
This commit changes how the MainNavigation behaves and moves it to a
Header component, removing the old header. Additionally it edits the
layouts so that the _Base layout is more useful and the Default and
Splash layout only impact what they actually change.

This move allows the header to be moved into the layout, removing it
from the individual pages.

These changes have the added benefit of fixing the header on the about
page.