Add content to index page
This adds content to the index page. It copies the mailchimp sign-up
form from the Barcelona website and adds a link to Twitter. It also
populates the CTA links for the ticket purchase and adds the location
building so it can be used for hotel planning purposes.

This uses the gatsby-source-rss-feed plugin to pull in the news at build
time so it also works when JavaScript is disabled.